Use `.data` instead of `.got.plt`
When libc has no symbols, to search the `main_arena` and `mp_` in the memory, we will need to find which page is possible to have those symbols first.

Before this commit, we used `.got.plt` to determine it, but it might fail in some cases. Using the address of `.data` should be a better solution.