Fix Arch CI: the load binary tests (#1400)
Before this commit we asserted whether the loaded binary in tests report
to find or not find debug symbols but this is irrelevant for the thing
we want to test there which is: pwndbg loading. What eventually cares is
whether Pwndbg got loaded and didn't raise an exception.

This commit fixes those tests so they should now work also on ArchLinux
CI and on all CI builds.

Additionally, it removes the `compile_binary` test utility function
which was redundant as we compile all test binaries via a makefile.