gdblib/symbol.py: remove symbol autofetch (#1494)
Removing this _autofetch: first of all, it never called gdb.execute
since it was commented out. Second of all, GDB seems to do symbol
autofetching by default or at least with my simple tests it was working.

Maybe a very old GDB version did not do that and that's we wanted to
have it in the first place.