Fix #855: Rust binaries debugging
This fixes #855 by using a workaround when we set language to C, fetch
symbols and set language back to Rust or auto.

Note that I also extended the list of symbols we fetched as... it seems
that the C language that is being set during some events may not have
e.g. `short` but may have a `short int` symbol. This is kinda weird, but
this is how it is. Idk if that's a GDB bug or what.

```
ipdb> gdb.execute('show language')
The current source language is "auto; currently c".
ipdb> gdb.lookup_type('short')
*** gdb.error: No type named short.
ipdb> gdb.lookup_type('int')
<gdb.Type object at 0x7fd35b550dc8>
ipdb> gdb.lookup_type('short int')
<gdb.Type object at 0x7fd35b550e68>
ipdb> gdb.lookup_type('long int')
<gdb.Type object at 0x7fd35b550e18>
```