Optimize coloring and fix a cache miss in generateColorFunction
This commit adds three small optimizations to the coloring layer:
1. It fixes a bug where we always triggered a cache miss on first call
to `generateColorFunctionInner` in `generateColorFunction` because we
recreated a lambda object on each call. This is described in more
details below.
2. The `none` function is now just assigned to `str` so that we don't
perform an unnecessary call
3. We now don't call `globals()` on each iteration of the loop in
`generateColorFunction` and instead cache the reference to the globals
dictionary in a fast lookup variable (default keyword argument). Note
that this means a `LOAD_FAST` is used for fetching it, instead of e.g.
`LOAD_GLOBAL` which would be the case if it were a global variable.

More details to 1.:

This commit fixes an embarassing and silly cache miss that we always hit
during calls like `pwndbg.color.memory.c.code` which really boils down to
calling the `ColorConfig.__getattr__` function.

The `ColorConfig.__getattr__` calls `generateColorFunction` which may call
`generateColorFunctionInner` multiple times, depending on thedesired
coloring.

The problem was that the `generateColorFunction` would always call the
`generateColorFunctionInner` with a "convert to string" function, which
was created as a local variable with the following code:

```
function = lambda x: str(x)
```

Since we always passed a new lambda to the `generateColorFunctionInner`
function, the memoization/caching layer always had a cache miss because
the hashing of a *new* lambda object always lead to a new hash, even if
the lambda seems to be the same thing.

Here are some benchmarks for just this 1) matter:

Before:
```
In [7]: %timeit pwndbg.color.memory.c.code('')
5.61 µs ± 135 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [8]: %timeit pwndbg.color.memory.c.code('')
5.65 µs ± 291 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [9]: %timeit pwndbg.color.memory.c.code('')
5.63 µs ± 28.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [10]: %timeit pwndbg.color.memory.c.code('')
5.77 µs ± 704 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:
```
In [1]: %timeit pwndbg.color.memory.c.code('')
3.68 µs ± 22.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [2]: %timeit pwndbg.color.memory.c.code('')
3.66 µs ± 31.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [3]: %timeit pwndbg.color.memory.c.code('')
3.7 µs ± 23.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [4]: %timeit pwndbg.color.memory.c.code('')
3.68 µs ± 24.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [5]: %timeit pwndbg.color.memory.c.code('')
3.63 µs ± 15 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```