feature: Add `killthreads` command (closes #1580)
This command allows the user to quickly kill multiple  threads by
specyfying their ids as arguments to this command. It also supports
the `--all` flag, which will kill overy thread except the currently
selected one. This is useful for use with the `checkpoint` command.

The killing is done by calling `pthread_exit(0)`.