Avoid downloading files from remote server in some cases
In some cases, we might not need to download the file if we can find the file locally, and this can be determined by the return value of the gdb command/API.

For example, the path we get from `pwndbg.gdblib.proc.exe` has different when gdb have loaded the file or not:

If we have executed `file /path/to/exe` before connecting to gdbserver:
```pycon
In [1]: pwndbg.gdblib.proc.exe
Out[1]: '/path/to/exe'
```
The output doesn't have "target:" prefix because gdb is able to find the file locally.

If we haven't executed `file /path/to/exe`, gdb can't find it locally, the output will be:
```pycon
In [1]: pwndbg.gdblib.proc.exe
Out[1]: 'target:/path/to/exe'
```

So if the "target:" prefix doesn't present when using `pwndbg.gdblib.file.get_file(pwndbg.gdblib.proc.exe)`, we don't need to download the file.
(The behavior of file path also occurred in `info files` and `info sharedlibrary` commands output.)

This commit aims to prevent downloading the file when the "target:" prefix doesn't present in the above cases.