feature: Add `killthreads` command (closes #1580) (#1581)
* feature: Add `killthreads` command (closes #1580)

This command allows the user to quickly kill multiple  threads by
specyfying their ids as arguments to this command. It also supports
the `--all` flag, which will kill overy thread except the currently
selected one. This is useful for use with the `checkpoint` command.

The killing is done by calling `pthread_exit(0)`.

* fix: try fixing building test binaries by enabling -lpthread

* fix: remove error message check when calling pthread_exit

Removed the message check, because the error messages difffer between
versions of GDB.

* fix: Improve UX of the killthreads command

Add an extended description of the command, some validation on the thread IDs
and suppress GDB output.

* fix: lint

* fix: put the multiline help text in the correct place

* tests: fix test failing due to a race condition when running in parrallel to other tests

Replaced asserts with loops which wait for a cundition to be met, so that the tests doesn't fail due to scheduling issues.

* tests: add more fixes for race conditions in test_killthreads

* fix: lint

* Update pwndbg/commands/killthreads.py

* tests: Wait for exactly three threads

---------

Co-authored-by: Disconnect3d <dominik.b.czarnota@gmail.com>