Optimize vis_heap_chunks command
This commit optimizes the vis_heap_chunks command by precalculating bin
labels instead of computing them on demand for each chunk.

For `vis 2000` command ran when debugging python3 shell, those changed
cut down the execution time from almost 20s to 5s.

The benchmark done is included in this commit in
profiling/benchmark_vis_heap_chunks/ so that it can be reproduced e.g.
to optimize the function further or to reproduce my results.