Add heuristic to add executable base in qemu-user
vmmap would try to add the executable to memory pages if the `info auxv`
command contained an address, but the memory maps would be accessed
recursively when trying to lookup the start of the ELF based on the
given address.

Since qemu doesn't provide memory map info, do a leap of faith and try
if the start of the page of the given address contains the ELF magic
header.

Since the program headers are more likely to be on the same page as the
ELF header than the program entrypoint, try both.