Stop using `CS_GRP_CALL` directly to determine what's a call
This commit introduces a new function (`is_call`) to `pwndbg.disasm`,
which centralizes the mechanism used to decide what instructions should
be considered calls and which ones shouldn't. Additionally, it uses this
new mechanism to flag the `bl` family of instructions on ARM as calls,
which, as of the time of writing, are not flagged as such by Capstone.