Unable to set param.value when encoding not UTF-8
Summary
-------
Prior to this change, when attempting to run pwndbg in Ubuntu 22.04.03
LTS docker container (without setting the code to UTF-8) the following
pwn.lib.config.Parameter(s) were unable to be set:
- chain-arrow-left
- chain-arrow-right
- banner-separator
- nearpc-branchmarker
- nearpc-prefix
- telescope-offset-separator
- telescope-repeating-marker
- backtrace-prefix
- code-prefix
- hexdump-ascii-block-separator

All of these params correspond to the UI for pwndbg. By bypassing the
set of param.value for each of these the pwndbg now runs on arm64
(inside a docker container thanks to QEMU).