Small fixes for RISC-V (#2172)
* Fix i386-32 syscall name printing

pwndbg-git from AUR shows hexadecimal constants in masm syntax
(e.g. 80h) for some reason (as if the option CS_OPT_SYNTAX_MASM was set).
This commit makes syscall name printing work regardless of hex syntax.

* riscv: Fix AssertionError on "jalr ra, ra, 0x252"

When the PC was on this instruction, the pwndbg context would not be
printed due to this AssertionError.

* riscv: Fix AssertionError on "c.jalr a5"

According to the specification, "C.JALR expands to jalr x1, 0(rs1)".