Fix terminal width to 80 columns in tests (#2444)
* Fix terminal width to 80 columns in tests

Set a `PWNDBG_IN_TEST`environment variable when running gdb in tests. Use the dimensions in `LINES` and `COLUMNS` when looking up the window size when that envvar is set.

This makes context output always be 80 columns wide which allows to compare to hardcoded output.

* Use `width` and `height` parameters for cmd window size

They are updated to the correct values when switching between tui and cli mode. So it's unnecessary to parse `info win`.

* Remove workarounds for terminal size in DEVELOPING

* Don't bother updating the env of the debuggee