Fix gdt command: require address argument
Before this commit, the `gdt` command incorrectly did not require the `address` argument, while it was required.

Because of that, the command crashed like this:

```
pwndbg> gdt
'gdt': Decode X86-64 GDT entries at address

See also:
* https://wiki.osdev.org/Global_Descriptor_Table
* https://wiki.osdev.org/GDT_Tutorial

Note:
In 64-bit mode, the Base and Limit values are ignored, each descriptor covers the entire linear address space regardless of what they are set to.
Exception occurred: gdt: int() argument must be a string, a bytes-like object or a real number, not 'NoneType' (<class 'TypeError'>)
For more info invoke `set exception-verbose on` and rerun the command
or debug it by yourself with `set exception-debugger on`
```