lldb: add support for attaching via `attach <pid|name>`
This commit adds support for the `attach <pid|name>` command so that it
can attach to a pid or full process name similarly to as in LLDB.

Please note that this LLDB command is a bit confusing.
It may seem it is an alias for `process attach` as in here:

```
(lldb) attach -n -w htop
error: 'process attach' doesn't take any arguments.
```

However, in practice it is not. It is an alias for `_regexp-attach`:

```
(lldb) help attach
Attach to process by ID or name.  Expects 'raw' input (see 'help raw-input'.)

Syntax: _regexp-attach <pid> | <process-name>

'attach' is an abbreviation for '_regexp-attach'
```

...which has its own problems:
1) it does not perform any regexp match in practice. passing `hto.*`,
   `hto?` or `hto[p]` will not attach to `htop`. One must pass the full process name like `htop`.
2) it can work without arguments, which is stupid and we should not
   support it?:

```
(lldb) _regexp-attach
There is a running process, detach from it and attach?: [Y/n] y
Process 56358 detached
Process 56358 stopped
* thread #1, queue = 'com.apple.main-thread', stop reason = signal SIGSTOP
    frame #0: 0x00007ff8118bedd2 libsystem_kernel.dylib`__select + 10
libsystem_kernel.dylib`__select:
->  0x7ff8118bedd2 <+10>: jae    0x7ff8118beddc ; <+20>
    0x7ff8118bedd4 <+12>: movq   %rax, %rdi
    0x7ff8118bedd7 <+15>: jmp    0x7ff8118b737b ; cerror
    0x7ff8118beddc <+20>: retq
```