Workaround GDB TUI ANSI escape sequence translation bug
Only resetting the foreground or background colors results in generating wrong colors in the TUI.

The workaround is to avoid the sequences that only reset the colors and replace them with a full reset sequence. It resets other styling attributes like bold too but it's better than having wrong colors.

Refs #2654