Switch to using `pwnlib` constants for flags.
`pwnlib` already exposes common constants in a platform-agnostic manner.
This was going to be a follow-up to the original flag formatting
support, but `pwnlib`'s functionality makes this trivial.

However, some platforms don't support all of the constants we might want
to associate with a specific function's flags. In this case, accessing
the constant from `pwnlib` will raise an `AttributeError`. The
`resolve_constants` helper function simply ignores the error to return
the constants that exist for the current platform.