exec_shellcode: add steps=N argument to prevent deadlocks
This adds a `steps=N` argument for
`pwndbg.aglib.shellcode.exec_shellcode`. When this is passed, the
underlying debugger will perform `stepi` N times instead of `continue`.

This helps prevent deadlocks e.g. when using `exec_shellcode` in Linux
kernel since `stepi N` is not interrupted by IRQ while `continue` can
be.

Of course this has its own issues, but I am not sure of a better
solution for now (and we can't just do `stepi 1000` until we hit an
expected PC address or something, because there may be loops).