Fill full height of context section in TUI mode
Add new `context-tui-adjust-height` config option to dynamically increase
the number of lines printed in `regs`, `disasm`, `code`, and `threads` sections
to fill the whole TUI window allocated for the section.

This ignores the other config values in place to limit the number of lines
for each individual section and interacts weirdly with the context history,
because the same entry could be added twice with longer section output
when switching from CLI to TUI mode and the context was already output
in CLI mode. That is a future TODO to replace the section data in the latest
history entry when applicable.