Add vmlinux command to load kernel syms from bzImage/vlinux
This commit implements the `vmlinux` command that loads kernel debug
symbols from a `bzImage` or `vmlinux` files. It works by loading it
through `vmlinux-to-elf` tool that needs to be in `$PATH` or provided as
`--tool <path>` argument.

I am not sure but for some reason it doesn't work for me that the tool
is in `$PATH` and I have to pass it as argument.

Then the `vmlinux` command extracts the kernel ELF from the kernel
object file and then loads it at appropriate kernel base address
detected by Pwndbg.

This then makes commands such as `kmod` to work correctly.

This is still WIP, but can be merged in current form.

It needs:
1. Vendoring of vmlinux-to-elf tool.
1. Caching of the resulting kernel ELF file as we currently re-create it
   on each call to `vmlinux` command.